db/src/requirements.txt
psycopg2-binary==2.9.1
pandas==1.3.5
